Creation of Sprint Bar
The Sprint Bar reflects the remaining stamina that the player has left to sprint with a sprite. The code in the Player Code is what determines how long the player can sprint for and how the player loses and regains stamina. The code in the SprintBar Code is what determines what sprite the Sprint Bar uses depending on how much stamina the player has.