Created Most of Inventory System
Added Item functionality such that when you pick up an item, it optionally disappears from the room and it gets added to your inventory.
Both the capacity of one slot (you can only hold a certain amount of an item) and the inventory capacity as a whole (you can only hold this many items) are functional.