Flashlight FIXED
The collision lines used in the flashlight can now properly detect objects that are specified in the code. The flashlight also stays in the most previous direction that the player is looking in. So if the player is moving North, then the flashlight is going to be looking North, even if the player stops moving.