Basic Enemy Movement Done
Created an Enemy_Parent Object. Created a Basic Enemy that can follow the player as long as the player is in range of the enemy. The enemy object uses the function move_towards_point() and inside the function is the player's position and the speed at which the enemy should be moving at. The enemy parent also has code that calculates the distance from the player every step. I plan on expanding on this in the future and implementing more diverse enemy movement patterns that are capable of using the level layout in order to better catch the player. Examples of this would be hiding behind walls and/or places that the player would have a hard time seeing unless they get close to it.