New Enemy Type In Progress
Currently working on a new enemy type that follows the player and then stops a certain distance away for 1 second, and then sprints at the player very quickly unless the player walks up to the enemy and shines their flashlight on them. I plan on making it so that the enemy stores the players position once it starts the wait time and also make it that this enemy is only capable of traveling in straight lines. This will be done by modifying the original enemy movement code and having it so that this enemy only detects the player's position in 5 second intervals.