Changes made to New Enemy and Tweaked Player Flashlight Size
The new enemy now only detects the player's position every 5 seconds, starting when they enter the scene. The enemy will go towards that position until it detects the players position after that, which it will then proceed toward that position.

Changed the size of the player's detection radius to better fit the games aspect ratio. Halved the length of the collision_lines, and got reduced the number of collision lines there were. This makes the detection radius more narrow as the outside 4 lines were commented out.