New Enemy Created
Created a new enemy type that is stationary but will make the player character move slower while in the enemy radius. Am currently working on a mechanic where if the enemy gets into the player's flashlight range, then the enemy will get stunned, and will also cause any enemy in their range to stop moving as well. This was accomplished using collision_circle_list to get all instances of enemies within the radius and then stopping their movement by stopping the move_towards_point line that is in the enemies step event.