Enemy Spawner Created and New Enemy Fixed
Fixed the newest enemy created where enemies inside its range will be stopped instead of continuing to move to the position they last marked.

Created an Object that handles spawning of enemies in the room. It uses delta_time to time when certain enemies spawn in the room. There is a random timer for how long it takes for another enemy of it's kind to spawn in the room. This was done so enemies will not spawn every frame for when they are capable of spawning and to add a bit of randomness to the spawning. The enemies also have a random range of where they can spawn.